+++ src/main/java/com/student/boot/common/annotation/Log.java
@@ -1,0 +1,23 @@
+package com.student.boot.common.annotation;
+
+import com.student.boot.common.enums.LogModuleEnum;
+
+import java.lang.annotation.*;
+
+/**
+ * 日志注解
+ *
+ * @author Ray
+ * @since 2024/6/25
+ */
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+@Documented
+public @interface Log {
+
+    String value() default "";
+
+    LogModuleEnum module()  ;
+
+
+}